CORE-1077: Rework `BRCryptoTimestamp` as `uint64_t`

+++ WalletKitCore/include/BRCryptoBase.h
@@ -77,8 +77,8 @@ extern uint64_t BLOCK_HEIGHT_UNBOUND_VALUE;
 
 #define BLOCK_NUMBER_UNKNOWN        (BLOCK_HEIGHT_UNBOUND)
 
-    /// The Timestamp (in the Unix epoch) of the last block processed in a sync.
-    typedef uint32_t BRCryptoTimestamp;
+    /// The number of seconds since the Unix epoch).
+    typedef uint64_t BRCryptoTimestamp;
 
 #define AS_CRYPTO_TIMESTAMP(unixSeconds)      ((BRCryptoTimestamp) (unixSeconds))
 #define NO_CRYPTO_TIMESTAMP                   (AS_CRYPTO_TIMESTAMP (0))
